fix: resolve 3 compile errors from VPS build
- main.rs: cast pos.cycles_held (u64) to u32 for TradeExit event fields
- daily_analyst.rs: add `..` wildcard to TradeEntry pattern (notional_usd unused)
- trade_log.rs: remove unused Duration import
- daily_analyst.rs: remove unused TradeLogger import

+++ src/main.rs
@@ -1404,8 +1404,8 @@ async fn close_paper_position(
             pnl_pct,
             r_multiple:       r_at_close,
             reason:           reason.to_string(),
-            cycles_held:      pos.cycles_held,
-            minutes_held:     pos.cycles_held / 2,
+            cycles_held:      pos.cycles_held as u32,
+            minutes_held:     (pos.cycles_held / 2) as u32,
             dca_count:        pos.dca_count,
             tranches_closed:  pos.tranches_closed,
         });
